+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Feat: Add Status code to http request (#1601)
+
 ## 5.1.0-beta.4
 
 * Update sentry-native to 0.4.11 (#1591)